libra-backendv2/requirements.txt
alembic==1.4.2
bcrypt==3.1.7
cffi==1.14.0
click==7.1.1
coloredlogs==14.0
Cython==0.29.15
cyvcf2==0.11.6
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
flask-marshmallow==0.11.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.1
humanfriendly==8.1
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
MarkupSafe==1.1.1
marshmallow==3.5.1
marshmallow-sqlalchemy==0.22.3
numpy==1.18.2
psycopg2==2.8.4
psycopg2-binary==2.8.4
pycparser==2.20
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
PyVCF==0.6.8
six==1.14.0
SQLAlchemy==1.3.15
Werkzeug==1.0.0
